+++ src/main/java/com/tiendis/tiendis/DAO/BancoDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.Banco;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BancoDAO extends CrudRepository<Banco, Long> {
+public interface BancoDAO extends CrudRepository<Banco, Integer> {
 }
