+++ src/main/java/com/tiendis/tiendis/DAO/CarritoDao.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.Carrito;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarritoDao extends CrudRepository<Carrito, Long> {
+public interface CarritoDao extends CrudRepository<Carrito, Integer> {
 }
