+++ src/main/java/com/tiendis/tiendis/DAO/CategoriaDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.Categoria;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoriaDAO extends CrudRepository<Categoria, Long> {
+public interface CategoriaDAO extends CrudRepository<Categoria, Integer> {
 }
