+++ src/main/java/com/tiendis/tiendis/DAO/ClienteDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.Cliente;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClienteDAO extends CrudRepository<Cliente, Long> {
+public interface ClienteDAO extends CrudRepository<Cliente, Integer> {
 }
