+++ src/main/java/com/tiendis/tiendis/DAO/CuentaBancariaDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.CuentaBancaria;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CuentaBancariaDAO extends CrudRepository<CuentaBancaria, Long> {
+public interface CuentaBancariaDAO extends CrudRepository<CuentaBancaria, Integer> {
 }
