+++ src/main/java/com/tiendis/tiendis/DAO/FormaPagoDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.FormaPago;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FormaPagoDAO extends CrudRepository<FormaPago, Long> {
+public interface FormaPagoDAO extends CrudRepository<FormaPago, Integer> {
 }
