+++ src/main/java/com/tiendis/tiendis/DAO/SubCategoriaDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.SubCategoria;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SubCategoriaDAO extends CrudRepository<SubCategoria, Long> {
+public interface SubCategoriaDAO extends CrudRepository<SubCategoria, Integer> {
 }
