+++ src/main/java/com/tiendis/tiendis/DAO/TipoClienteDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.TipoCliente;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TipoClienteDAO extends CrudRepository<TipoCliente, Long> {
+public interface TipoClienteDAO extends CrudRepository<TipoCliente, Integer> {
 }
