+++ src/main/java/com/tiendis/tiendis/DAO/TipoComercioDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.TipoComercio;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TipoComercioDAO extends CrudRepository<TipoComercio, Long> {
+public interface TipoComercioDAO extends CrudRepository<TipoComercio, Integer> {
 }
