+++ src/main/java/com/tiendis/tiendis/DAO/TipoDocumDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.TipoDocum;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TipoDocumDAO extends CrudRepository<TipoDocum, Long> {
+public interface TipoDocumDAO extends CrudRepository<TipoDocum, Integer> {
 }
