+++ src/main/java/com/tiendis/tiendis/service/CarritoService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.Carrito;
 
-public interface CarritoService extends GenericServiceApi<Carrito, Long> {
+public interface CarritoService extends GenericServiceApi<Carrito, Integer> {
 }
