+++ src/main/java/com/tiendis/tiendis/service/CategoriaService.java
@@ -3,6 +3,6 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.Categoria;
 
-public interface CategoriaService extends GenericServiceApi<Categoria, Long> {
+public interface CategoriaService extends GenericServiceApi<Categoria, Integer> {
 
 }
