+++ src/main/java/com/tiendis/tiendis/service/CuentaBancariaService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.CuentaBancaria;
 
-public interface CuentaBancariaService extends GenericServiceApi<CuentaBancaria, Long> {
+public interface CuentaBancariaService extends GenericServiceApi<CuentaBancaria, Integer> {
 }
