+++ src/main/java/com/tiendis/tiendis/service/EstadoCarritoService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.EstadoCarrito;
 
-public interface EstadoCarritoService extends GenericServiceApi<EstadoCarrito, Long> {
+public interface EstadoCarritoService extends GenericServiceApi<EstadoCarrito, Integer> {
 }
