+++ src/main/java/com/tiendis/tiendis/service/ProductoService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.Producto;
 
-public interface ProductoService extends GenericServiceApi<Producto, Long> {
+public interface ProductoService extends GenericServiceApi<Producto, Integer> {
 }
