+++ src/main/java/com/tiendis/tiendis/service/ProveedorService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.Proveedor;
 
-public interface ProveedorService extends GenericServiceApi<Proveedor, Long> {
+public interface ProveedorService extends GenericServiceApi<Proveedor, Integer> {
 }
