+++ src/main/java/com/tiendis/tiendis/service/TipoComercioService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.TipoComercio;
 
-public interface TipoComercioService extends GenericServiceApi <TipoComercio, Long> {
+public interface TipoComercioService extends GenericServiceApi <TipoComercio, Integer> {
 }
